move ethereum signer code to its own directory
