+++ packages/signers/src/index.ts
@@ -1,6 +1,6 @@
 import { Connector, SessionWallet, ActionWallet } from "./interfaces.js"
-import { MetaMaskEthereumConnector, MetaMaskEthereumSessionWallet } from "./metamask_web_wallet.js"
-import { WalletConnectWebWalletConnector } from "./walletconnect_web_wallet.js"
+import { MetaMaskEthereumConnector, MetaMaskEthereumSessionWallet } from "./ethereum/metamask_web_wallet.js"
+import { WalletConnectWebWalletConnector } from "./ethereum/walletconnect_web_wallet.js"
 
 export { ethersBlockToCanvasBlock, EthereumBlockProvider } from "./providers.js"
 export {
